Support navigation between index and guideline pages
In order for page navigation to work, scaffold has been moved out of
main and into the individual page widgets. Navigation has been added
to go from an index page to a specific guideline page.

InstantPageRoute is a version of MaterialPageRoute without a transition
animation. It is not currently used and may be removed if the decision
to stay with the material transition is finalised.