Switch to named navigation routes and refactor main widgets
Navigation is now via named routes. onGenerateRoute is used to parse
arguments to be passed into the route (for guideline display). The
home widget is now Index which generates a list of guidelines similar
to GuidelineList which it replaces. This index reads in available
guidelines from index.json in the guidelines folder. Have also merged
the GuidelineSection class into Guideline (though I'm not sure this
is better.