Major rework of the guideline widget
Firstly, the guideline data is now read into a GuidelineData object
and stored as the state of Guideline. Guideline itself now uses
FutureBuilder to handle the async data loading of GuidelineData.
Lastly, The different sections of the Guideline widget are again moved
out into their own objects. This was the initial design but the
implementation is much better now.

The result is a cleaner set of widgets and also this should set the
stage for allowing us to switch between private and commercial licence
views.