Basic styling of the app
Provide some styling for the app. The styling is fairly simple but
looks ok. Dart's extension mechanism is used to add custom properties
to ThemeData so all style can be stored in style.dart (other than
layout information which is generally in each widget). style.dart needs
to be included in all files that use any of those custom properties.