Add bottom navigation bar
Added the ability to navigate pages using a bottom navigation bar.
This is potentially not how it is designed to be used as it appears
to be mostly for replacing the body of the page rather than the entire
view. To achieve this functionality we essentially rebuild the bottom
nav bar every time we navigate pages. Also to control the view stack
we use replaceRoute rather than pushRoute and we use WillPopScope to
catch back button presses on the favourites and information routes and
move us back to the index page in those instances.

Overall everything appears to work as intended but this unusual usage
of the bottom nav bar may end up introducing some subtle bugs.