Add support for favourites and tweak navigation control
Added support for marking certain guidelines as a favourite. These
can then be seen in isolation in the Favourites view. Although Index
and favourites share many similarities, Index and Favourites are coded
as separate widgets for now as the display of Index and Favourites is
likely to change in the future. Once the overall design has stabilised
it might make sense to merge them into the same class with a togglable
option to show all guidelines or just the favourites.

Individual Guideline views don't show whether the displayed guideline
is a favourite or not. This should be implemented with the option to
toggle favourite state from within the Guideline view.

This commit also has changes to the way bottom navigation works. The
issue is that bottom navigation doesn't appear to be designed for
page navigation but rather to just change displayed content in the
master view. This has tweaked things so the leading back arrow in the
header no longer shows in all view. It could be enabled in the
Guideline view if needed but it looks better without. It may make sense
to eventually restructure the whole app so that all views are at a
single navigation route.