Rework of the guideline display view
A GuidelineContainer is now used to wrap the Guideline display. It
uses a TabBar to control viewing the private or commercial standard.

Although a relatively small change to the UI, it is a fairly large
rework to the relevant code.

Also implemented changes suggested by flutter analyze in this commit.