Convert the main index page to be an expansion list
Guidelines in the index page are now under categories and grouped. The
current implementation is messy with heavily nested code. It should be
cleaned up by converting to more separate widgets but the entire
guidelines setup is likely to be changed to a sub-folder and HTML based
one in the future so I will defer cleaning up until those plans are
finalised.