Begin work to refactor the main views
MainWindow is now defined as the parent widget of the index,
favourites and information view. It is set up as a provider to allow
child widgets to access and modify some of its data.

The main app bar and main nav bar are now separate widgets which are
consumers of the data provided by the main window.

The index, favourites and information views are not yet implemented.