Set up IndexModel as a provider of the guidelines index data
The guideline index is now made available to all children of the main
window. The structure of the map has also been changed. Each entry
represents a guideline and data is stored within the map as to the
name, categories and group of the guideline. The status as favourite
or not will also be stored within the map.

A new tool, create_index is included which replaces generate_index
for building the new index file.

Old files will need to be tidied up when the rework is complete.