Complete most work on the index page
The index page now displays correctly again with support for
favourites and expansion panels. The header is now also clickable to
expand an expansion panel.

The new implementation uses provider rather than FutureBuilder and
setState. It makes the architecture a little bit cleaner. For now
the preferences Consumer is relatively high up the widget tree. It
could be pushed further down if needed.

At the moment the only thing that doesn't work is moving to a guideline
page as the changes to how guideline data is handles is not worked into
the guideline views yet.