Reconnect navigation to the guideline window
The navigation is a bit cleaner now. By passing a route to the nav bar
it can determine how to display (highlighting a page or not) based on
which window it is being displayed in. Moving from the guideline route
back to the main window is just a pop() command now.

Display of the guideline still needs to be re-introduced.