Re-enable display of guideline data
This is still mostly implemented via the existing widgets within the
old widget subfolder. GuidelineContainer is replaced with
GuidelineWindow which uses the same app bar and nav bar as the other
views. GuidelineData is replaced with GuidelineModel. The new
implementation is a little cleaner under the hood.

The favourites button is moved out of the app bar into a floating
action button.