Complete the rework of the main views
The views are now split across two view - main and guideline. The main
view can now display the index, favourites or information pages. The
guideline view is a rework of the guideline container though the actual
guideline display code is unchanged from the original implementation.

This rework has a number of benefits:
 - Index and preferences loading is all done in one spot as a provider
   model
 - The data from the loaded index or preferences is available anywhere
   within the widget hierachy
 - By placing the three main view within the main window we avoid
   issues with managing the back button and scope
 - The overall code layout is cleaner with less repitition, less
   nesting and a simpler code-path

In terms of the user facing changes, the app is almost identical with
the following excpetions:
 - The text of the expandable header is now clickable
 - The favourite button on the guideline view is now a floating action
   button