Begin re-implementation of index page
Have started refactoring and rewritting code for the
ExpansionPanelList. The panels now work though the context of the
panels needs to be implemented.

Through this refactor the data model for the index will be changed to
be a little cleaner. For now the index data is hard-coded into the
widget.