Updated the index page
The index page now displays correctly again, loading data from the
index model as expected. Favouriting is again working on the index
page. The page still needs styling and needs to load the guideline
page when an item is clicked.

Changes have been made to the how the index model works and as such the
old version of then index and favourites page no longer works.