Modified how appbar title is specified
The old implementation work for the views with a static title (Index,
Favourites and Information), but not so well for the guideline view
where the title was dynamic.

The new implementation works for all views but triggers an extra,
unnecessary page load for the static title pages. This could be
partially reverted in the future so static titles use the old method
for whilst dynamic titles use the new method.

Additionally, some of the old implementation files have been removed
as the new implementation is now  mostly complete.